requirements.txt
SQLAlchemy
mysqlclient
mysql-connector-python
pycryptodome
tornado
redis
nltk
beautifulsoup4
wordcloud
algorithmia
